requirements.txt
customtkinter==5.2.2
dacite==1.9.2
pillow==10.4.0
pygame==2.6.1
pyinstaller==6.14.2
reportlab==4.4.3
